feat: put button availability state handler into a separate hook
