+++ src/Hooks/index.ts
@@ -5,7 +5,7 @@ export { default as useMainButton } from "./useMainButton/useMainButton";
 export { default as useRandomCards } from "./useRandomCards/useRandomCards";
 export { default as useLowBalancePopup } from "./useLowBalancePopup/useLowBalancePopup";
 export { default as useShareBotUrl } from "./useShareBotUrl/useShareBotUrl";
-export { default as useMainButtonTextAndHandler } from "./useMainButtonTextAndHandler/useMainButtonTextAndHandler";
+export { default as useMainButtonHandler } from "./useMainButtonHandler/useMainButtonHandler";
 export { default as useSaveSpreadState } from "./useSaveSpreadState/useSaveSpreadState";
 export { default as useNavigation } from "./useNavigation/useNavigation";
 export { default as useRewardPopup } from "./useRewardPopup/useRewardPopup";
@@ -13,3 +13,4 @@ export { default as useInfoPopup } from "./useInfoPopup/useInfoPopup";
 export { default as useErrorPopup } from "./useErrorPopup/useErrorPopup";
 export { default as useSettingsButton } from "./useSettingsButton/useSettingsButton";
 export { default as useReadings } from "./useReadings/useReadings";
+export { default as useMainButtonState } from "./useMainButtonState/useMainButtonState";
